solved bugs
